refactor: 코드 분리, 디렉토리 추가
